Change the way nick/username/realname are set
Rather than have properties, setting which does nothing, we now
have readonly properties for nickname/username/realname, and a single
setter method, which may be called prior to the session connecting.
That way we won't have people setting the properties, and having the
properties then not reflecting the state of the connection!